(feat): se modificó Brick para que funcione como model
-Ahora "Brick" ya no deriva de MonoBehavior
-Se creó un nuevo script "BrickBehavior" el cual sí deriva de Monobehavior.
-Brick ahora funciona como un Model.
-Se modificaron todos los scripts necesarios para que todo siga funcionando igual.
