+++ Unity/Assets/Clases/Scenes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6fd04010852edab44a9c3bb771f9d428
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
