+++ Unity/Assets/Clases/Scenes/Game.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e85247601291446499f81518720df6cb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
