+++ Unity/Assets/Clases/Scripts/Interfaces/IInteractable.cs
@@ -2,4 +2,6 @@
 public interface IInteractable
 {
     void Interact(PlayerController player);
+    void OnPickedUp(PlayerController player);
+    void OnDropped();
 }
